stage1.js and stage1cards.js done for the most part - updated Navbar in order to properly link the stage 1 (I removed the space between stage and 1.  Also, kept old stage1.js file in folder for reference - will delete later
