+++ client/src/Components/Navbar/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import "./style.css";
 
 function Navbar(props) {
-    const links = ["Login", "Main Page", "Store", "Stage 1", "Boss", "Defeat", "Victory"];
+    const links = ["Login", "Main Page", "Store", "Stage1", "Boss", "Defeat", "Victory"];
     const navLinks = links.map(link => {
         return (
             <li><a href={"/" + link}>{link}</a></li>
